Update example nest_1.nml for new HYCOM
NOTE: The fill value for the global HYCOM data is now -30000. However, the u and v velocities as well as temperature and salinity require a scaling factor of 0.001, which seems to be automatically applied on download. Therefore, the correct fill value to use for now is -30. Once you have run getdata to download the nest files, you need to replace this with the cms fill value of 1.2676506E30 for running cms.

NOTE 2: I took out the 'orthogrid' option as this doesn't seem to be mentioned anywhere in the user guide?