Update comments output.f90
In info that is printed to the screen, replaced the line "total number of release locations" with "total number of release events", as this value is actually the number of lines in the release file (i.e. no. of release locations * no. of release dates), therefore "number of release locations" could be confusing.