Bug fix mod_getdata.f90 re I/O error
This bug fix relates to an issue where getdata would only download a certain number of files before crashing with an 'I/O failure' error and having to be restarted.

Feedback from Chris Edsall at U. Bristol Advanced Computing Research Centre:
"It seems that something in the network path takes exception to the number of open TCP connections and sends a TCP reset. The reason there are a number of open connections is each time through the loop there is an extra NetCDF open (this can be seen by adding a line to print out ncId in each of nc_open and nc_close in mod_nciorw.f90)."

A call to nc_open at line 1543 was therefore removed. This seems to fix the issue.